+++ backend/team6/src/main/java/programmers/team6/domain/admin/repository/VacationInfoLogSearchRepository.java
@@ -1,4 +1,4 @@
-package programmers.team6.domain.admin.service;
+package programmers.team6.domain.admin.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
