+++ backend/team6/src/main/java/programmers/team6/domain/admin/service/MemberReader.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
 import programmers.team6.domain.admin.dto.VacationStatisticsRequest;
+import programmers.team6.domain.admin.repository.VacationInfoLogSearchRepository;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.member.repository.MemberSearchRepository;
